refactor: rename the service store module after RaftStore
# Summary

`service/src/store/store.rs` is now `raft_store.rs`, dropping the last
`#[allow(clippy::module_inception)]` in the server crates.

# Details

The directory earns its place: it also holds `meta_raft_log` and
`meta_raft_state_machine`. Only the file was misnamed, so it now says
what it defines. Every caller already reached the type through the
`pub use` in `mod.rs`, so no import changed.
